+++ Assets/Resources/ScriptableObjects/UpgradeScriptable.cs
@@ -6,8 +6,8 @@ using NaughtyAttributes;
 [CreateAssetMenu( fileName = "Upgrade", menuName = "ScriptableObjects/Upgrade" )]
 public class UpgradeScriptable : ScriptableObject
 {
-	[SerializeField] private float damageUpgrade;
-	public float DamageUpgrade { get => damageUpgrade; set => damageUpgrade = value; }
+	[SerializeField] private int damageUpgrade;
+	public int DamageUpgrade { get => damageUpgrade; set => damageUpgrade = value; }
 	[SerializeField] private float hitBoxUpgrade;
 	public float HitBoxUpgrade { get => hitBoxUpgrade; set => hitBoxUpgrade = value; }
 	[SerializeField] private float attackSpeedUpgrade;
