+++ Assets/Scripts/Player/AbilityFunctions/Abilities/Ability.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public abstract class Ability : MonoBehaviour ,IAbility
 {
 	protected float coolDown;
-	protected float damage;
+	protected int damage;
 	protected EdgeCollider2D shape;
 	protected float distance;
 	protected Vector2 boxSize;
@@ -27,7 +27,7 @@ public abstract class Ability : MonoBehaviour ,IAbility
 	public Transform CastFromPoint { get; set; }
 	public float Angle { get; set; }
 	public float CoolDown { get => coolDown; set => coolDown = value; }
-	public float Damage { get => damage; set => damage = value; }
+	public int Damage { get => damage; set => damage = value; }
 	public float CritChance { get => critChance; set => critChance = value; }
 	public Vector2 BoxSize { get => boxSize; set => boxSize = value; }
 	public GameObject CastObject { get => castObject; set => castObject = value; }
